+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 ### What needs to be done
 
 * sqlite3 Chinese display problem
-* functions: changing default currency, adding new currency
+* functions: adding new currency, delete currency
+* "/view" total balances (month/year view, whole)
 
 ### What has been done in v0.1
+
+* 2014-09-16  
+  fixed the bug that "/view" cost column only displays integers  
+  shows "/view" in date descending order  
+  "/view" total rows show monthly income, expenses and balance  
+  added the function that can change default currency
 
 * 2014-09-14  
   add function that can change default number of rows in one db insertion